Move GcsObjectVersion class to a separate file.
This is part of the work to break the testbench script into smaller
pieces.
